+++ host-software/keyplus/usb_ids.py
@@ -5,6 +5,8 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+from enum import Enum
+
 ###############################################################################
 #                                   classes                                   #
 ###############################################################################
@@ -26,20 +28,35 @@ class USBKeyplusKeyboardInfo(USBDeviceInfo):
         self.interface = interface
 
 class USBBootloaderInfo(USBDeviceInfo):
-    def __init__(self, vid, pid, description):
+    def __init__(self, vid, pid, bootloader, description):
         super(USBBootloaderInfo, self).__init__(
             vid, pid, description
         )
+        self.bootloader = bootloader
+
 
 ###############################################################################
 #                                  constants                                  #
 ###############################################################################
+
+class BootloaderType(Enum):
+    XUSB_BOOT = 0
+    NRF24LU1P_512 = 1
+    NRF24LU1P_FACTORY = 2
+    KP_BOOT_32U4 = 3
+
 
 KEYPLUS_USB_IDS = {
     (0x6666, 0x1111): USBKeyplusKeyboardInfo(
         vid = 0x6666,
         pid = 0x1111,
         description = "keyplus keyboard xmega (prototype)",
+    ),
+
+    (0x6666, 0x2222): USBKeyplusKeyboardInfo(
+        vid = 0x6666,
+        pid = 0x2222,
+        description = "keyplus keyboard atmega32u4 (prototype)",
     ),
 
     (0x6666, 0x3333): USBKeyplusKeyboardInfo(
@@ -65,24 +82,36 @@ BOOTLOADER_USB_IDS = {
     (0x6666, 0xB007): USBBootloaderInfo(
         vid = 0x6666,
         pid = 0xB007,
+        bootloader = BootloaderType.XUSB_BOOT,
         description = "xusb boot (prototype id)",
     ),
 
     (0x1209, 0xBB01): USBBootloaderInfo(
         vid = 0x1209,
         pid = 0xBB01,
+        bootloader = BootloaderType.XUSB_BOOT,
         description = "keyplus xusb boot bootloader",
     ),
 
     (0x1209, 0xBB03): USBBootloaderInfo(
         vid = 0x1209,
         pid = 0xBB03,
+        bootloader = BootloaderType.NRF24LU1P_512,
         description = "keyplus nrf24lu1p-512 bootloader",
+    ),
+
+    # (0x1209, 0xBB08): USBBootloaderInfo(
+    (0x6666, 0x9999): USBBootloaderInfo(
+        vid = 0x6666,
+        pid = 0x9999,
+        bootloader = BootloaderType.KP_BOOT_32U4,
+        description = "kp_boot_32u4 bootloader",
     ),
 
     (0x1915, 0x0101): USBBootloaderInfo(
         vid = 0x1915,
         pid = 0x0101,
+        bootloader = BootloaderType.NRF24LU1P_FACTORY,
         description = "Nordic nRF24LU1+ factory bootloader",
     ),
 }